airflow/requirements.txt
faker
tqdm
minio
python-dotenv